Modified a few questions that students stumbled on.
Q2.2: The second part of this question can be hard to understand
if you don't recognize the dynamical importance of type II vs
type I neurons. Having noted the obvious qualitative diff,
some students think they're missing something and make stray
observations in response.

Q2.4: Some students were unclear what was being asked of them
with regards to the mathematical manipulation in 2.4.

Q3.1: The ordering language confused a lot of students, and many
would give "orderings" and explanations that were at odds.

Q3.3: The actual order of unit activation is not necessarily
the same as the order of correct unit activation, which is what
the question is really about. Students would thus report wrong
and or highly detailed info about transients rather than the main
thing.